ref: moving repository interfaces to application layer
