+++ src/app/features/user_management/application/use_cases/read/get_all_users_case.py
@@ -2,7 +2,7 @@ from typing import List
 
 from src.app.features.user_management.application.dtos.user_dto import UserResponse
 from src.app.features.user_management.application.dtos.user_dto_mapper import map_entity_to_dto_user
-from src.app.features.user_management.domain.repositories.user_repository import UserRepository
+from src.app.features.user_management.application.repository.user_repository import UserRepository
 from src.shared.application.base_use_case import BaseUseCase
 from src.shared.utils.log_util import log
