+++ src/app/features/user_management/application/use_cases/read/get_users_by_role_case.py
@@ -4,7 +4,7 @@ from src.app.features.user_management.application.dtos.user_dto import UserRespo
 from src.app.features.user_management.application.exceptions.user_exceptions import InvalidRoleException
 from src.app.features.user_management.application.dtos.user_dto_mapper import map_entity_to_dto_user
 from src.app.features.user_management.domain.entities.user_enums import UserRole
-from src.app.features.user_management.domain.repositories.user_repository import UserRepository
+from src.app.features.user_management.application.repository.user_repository import UserRepository
 from src.shared.application.base_use_case import BaseUseCase
 from src.shared.utils.log_util import log
