+++ src/app/features/user_management/application/use_cases/write/delete_user_case.py
@@ -1,5 +1,5 @@
 from src.app.features.user_management.application.exceptions.user_exceptions import UserDoesNotExistException
-from src.app.features.user_management.domain.repositories.user_repository import UserRepository
+from src.app.features.user_management.application.repository.user_repository import UserRepository
 from src.app.features.user_management.domain.value_objects.user_id import UserId
 from src.shared.utils.log_util import log
 from src.shared.utils.uuid_util import convert_str_to_uuid
