+++ src/app/features/user_management/application/use_cases/write/update_user_case.py
@@ -1,7 +1,7 @@
 from src.app.features.user_management.application.dtos.user_dto import UpdateUserRequest, UserResponse
 from src.app.features.user_management.application.exceptions.user_exceptions import UserDoesNotExistException
 from src.app.features.user_management.application.dtos.user_dto_mapper import map_entity_to_dto_user
-from src.app.features.user_management.domain.repositories.user_repository import UserRepository
+from src.app.features.user_management.application.repository.user_repository import UserRepository
 from src.app.features.user_management.domain.value_objects.user_id import UserId
 from src.shared.utils.log_util import log
 from src.shared.utils.uuid_util import convert_str_to_uuid
